refactor: move duplicated types in app and asset-utils to shared
