+++ packages/app/package.json
@@ -43,6 +43,7 @@
     "zustand": "^5.0.7"
   },
   "devDependencies": {
+    "@depl/shared": "workspace:^",
     "@eslint/js": "^9.11.1",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/lodash.merge": "^4.6.9",
