+++ packages/asset-utils/package.json
@@ -9,6 +9,7 @@
     "husky:pre-commit": "pretty-quick --staged --pattern \"packages/asset-utils/**/*.{js,ts,jsx,tsx,json}\""
   },
   "devDependencies": {
+    "@depl/shared": "workspace:^",
     "@types/node": "^22.7.5",
     "@types/unzipper": "^0.10.10",
     "prettier": "^3.3.3",
